Provide support for a warning threshold
Introduces key->value kwargs for parsing out two new optional options, a
warning threshold and a path on the filesystem to touch when we get a
warning. This way you can monitor that file's modification time and
automatically begin capturing CPU profiles or other diagnostic
information when we are approaching OOM behavior (or just general bad GC
behavior).