Portable builds and lots of testing.
Refactors the source to be portable, compiles with portable compiler
(supports clang and gcc now) and includes a lot of dockerized testing
to assure that it works on different JVMs and libcs.

One interesting part of this is that while compilation requires a JDK,
the generated .so does not link to any JVM libraries. As long as the
JVMTI interface remains stable (tested to be true on at least Java 8 and
Java 11) jvmquake should continue functioning.

When new Java's arrive we can just add them to our test matrix!